# swagger.yaml
openapi: 3.0.4
info:
  title: Taylor's Touch API
  version: 1.0.0
  description: API for managing users, clients, massages, and bookings
servers:
  - url: http://localhost:3000
paths:
...
  /bookings:
    get:
      summary: Get all bookings
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '500':
          description: Server error fetching bookings
    post:
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_id
                - massage_id
                - booking_date
                - start_time
                - end_time
                - status
              properties:
                client_id:
                  type: integer
                massage_id:
                  type: integer
                booking_date:
                  type: string
                  format: date
                start_time:
                  type: string
                  format: time
                end_time:
                  type: string
                  format: time
                status:
                  type: string
                  enum: [scheduled, completed, cancelled]
                notes:
                  type: string
            example:
              client_id: 1
              massage_id: 1
              booking_date: '2026-04-20'
              start_time: '10:00'
              end_time: '10:30'
              status: 'scheduled'
              notes: 'may need a little extra attention on the neck area'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid input
        '500':
          description: Server error creating booking
  /bookings/{id}:
    get:
      summary: Get one booking by unique ID
      parameters:
        - name: id
          in: path
          required: true
          description: The booking ID
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
        '500':
          description: Server error fetching booking
    put:
      summary: Update an existing booking
      parameters:
        - name: id
          in: path
          required: true
          description: The booking ID to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_id
                - massage_id
                - booking_date
                - start_time
                - end_time
                - status
              properties:
                client_id:
                  type: integer
                massage_id:
                  type: integer
                booking_date:
                  type: string
                  format: date
                start_time:
                  type: string
                  format: time
                end_time:
                  type: string
                  format: time
                status:
                  type: string
                  enum: [scheduled, completed, cancelled]
                notes:
                  type: string
            example:
              client_id: 1
              massage_id: 1
              booking_date: '2026-04-25'
              start_time: '13:00'
              end_time: '14:00'
              status: 'scheduled'
              notes: 'may need a little extra attention on the neck area'
      responses:
        '200':
          description: Booking successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid input
        '404':
          description: Booking not found
        '500':
          description: Server error updating booking
    delete:
      summary: Delete an existing booking
      parameters: 
        - name: id
          in: path
          required: true
          description: The booking ID to delete
          schema:
            type: integer
      responses:
        '204':
          description: Booking successfully deleted
        '404':
          description: Booking not found
        '500':
          description: Server error deleting booking
components:
  schemas:
...
    Booking:
      type: object
      properties:
        booking_id:
          type: integer
        user_id:
          type: integer
        massage_id:
          type: integer
        booking_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time